requirements.txt
astroid==2.6.5
certifi==2021.5.30
click==8.0.1
cloudinary==1.26.0
colorama==0.4.4
dnspython==1.16.0
Flask==2.0.1
Flask-Cors==3.0.10
gunicorn==20.1.0
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
ps==0.0.2
pylint==2.9.6
pymongo==3.12.0
python-dotenv==0.19.0
six==1.16.0
toml==0.10.2
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
